rewrote to return index instead of value

+++ utils/ArrayMethods/ArrayMethods.h
@@ -7,6 +7,6 @@ int arr_average(int arr[], int n);
 int getClosestInt(int val1, int val2, int target);
 float getClosestFloat(float val1, float val2, float target);
 int arr_find_closest_int(int arr[], int n, int target);
-float arr_find_closest_float(float arr[], int n, float target);
+int arr_find_closest_float(float arr[], int n, float target);
 
 #endif
